Refactored Stack Trace to be hosted by Interpreter
